backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
minio
python-multipart
python-jose[cryptography]
passlib[bcrypt]